+++ src/main/java/com/recipe/app/src/recipe/infra/youtube/YoutubeRecipeCustomRepository.java
@@ -10,7 +10,7 @@ public interface YoutubeRecipeCustomRepository {
 
     Long countByKeyword(String keyword);
 
-    List<YoutubeRecipe> findByKeywordLimit(String keyword, int size);
+    List<YoutubeRecipe> findByKeyword(String keyword);
 
     List<YoutubeRecipe> findByKeywordLimitOrderByPostDateDesc(String keyword, Long lastYoutubeRecipeId, LocalDate lastYoutubeRecipePostDate, int size);
